REFACTOR: sse endpoint refactoring(#14)
