+++ nadeulAI_SSE/src/constants/signals.py
